Cast sample duration to Float

+++ sm-audiokit/Audio/Sample.swift
@@ -10,9 +10,9 @@ import Foundation
 struct Sample {
     var id: String
     var url: URL
-    var duration: Float64    
+    var duration: Float
 
-    init(id: String, url: URL, duration: Float64) {
+    init(id: String, url: URL, duration: Float) {
         self.id = id
         self.url = url
         self.duration = duration
